# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: back-work API
  version: 0.0.1-SNAPSHOT
  description: |
    Documentation OpenAPI generee a partir des controllers detectes dans le projet.
servers:
  - url: http://localhost:8080

paths:
  /games:
    get:
      summary: Lister les jeux (avec filtres optionnels)
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Filtrer par nom (egalite, insensitive case)
        - name: genre
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/GameGenre"
          description: Filtrer si le jeu contient ce genre
        - name: minPrice
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
          description: Prix minimum inclus
        - name: maxPrice
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
          description: Prix maximum inclus
      responses:
        "200":
          description: Liste de jeux
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      summary: Creer un jeu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameCreate"
      responses:
        "201":
          description: Jeu cree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "500":
          description: Erreur interne
    delete:
      summary: Purger les jeux (games + tables de jointure)
      responses:
        "204":
          description: Purge effectuee

  /games/{id}:
    get:
      summary: Recuperer un jeu par id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Jeu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "500":
          description: Erreur interne

  /users:
    get:
      summary: Lister les utilisateurs
      responses:
        "200":
          description: Liste d'utilisateurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      summary: Creer un utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Utilisateur cree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "500":
          description: Erreur interne
    delete:
      summary: Purger les utilisateurs (users + table de jointure)
      responses:
        "204":
          description: Purge effectuee

  /users/{id}:
    get:
      summary: Recuperer un utilisateur par id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Utilisateur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "500":
          description: Erreur interne
    put:
      summary: Mettre a jour un utilisateur
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Utilisateur mis a jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "500":
          description: Erreur interne
    delete:
      summary: Supprimer un utilisateur par id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Utilisateur supprime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "500":
          description: Erreur interne

components:
  schemas:
    Game:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: integer
          format: int64
          nullable: true
        genre:
          type: array
          items:
            $ref: "#/components/schemas/GameGenre"
          uniqueItems: true
          nullable: true
        makingTime:
          type: string
          format: date-time
          nullable: true
      required:
        - name

    GameCreate:
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
          format: int64
          nullable: true
        genre:
          type: array
          items:
            $ref: "#/components/schemas/GameGenre"
          uniqueItems: true
          nullable: true
        makingTime:
          type: string
          format: date-time
          nullable: true
      required:
        - name

    GameGenre:
      type: string
      enum:
        - action
        - romance
        - thriller
        - horror
        - multiplayer
        - singleplayer

    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        isAdmin:
          type: boolean
        recommendedGames:
          type: array
          items:
            $ref: "#/components/schemas/Game"
          uniqueItems: true
          nullable: true
      required:
        - username
        - email
        - isAdmin

    UserCreate:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        isAdmin:
          type: boolean
        recommendedGames:
          type: array
          items:
            $ref: "#/components/schemas/Game"
          uniqueItems: true
          nullable: true
      required:
        - username
        - email
        - isAdmin

    UserUpdate:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        isAdmin:
          type: boolean
        recommendedGames:
          type: array
          items:
            $ref: "#/components/schemas/Game"
          uniqueItems: true
          nullable: true
      required:
        - username
        - email
        - isAdmin